pyproject.toml
[project]
name = "api-fut-pay-manager"
version = "0.1.0"
description = "API REST para gerenciamento de caixa de futsal"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "motor>=3.0",
    "pyjwt>=2.8",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "B008",  # function call in default arg — standard FastAPI pattern (Depends, Query)
    "UP046", # Generic subclass — Pydantic doesn't fully support PEP 695 type params
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
